Update ChangeUsername/ChangeAccountname to their new Views.
